任务1: 1.完成tebi存储服务的注册，并完成存储桶创建，存储区域选择物理位置离中国最近的两个区域。 2.创建存储桶密钥，需要给s3的所有权限，然后完成tebi工具类，其中需要使用s3协议。工具类需要包含上传文件，下载文件，获取文件上传预签名等基本接口
任务2：
1.用户类增加头像字段，增加用户注册接口，注册的时候需要传入手机号邮箱等字段，注册的时候需要检验手机号，邮箱是否唯一，并且注册的时候需要上传头像
2.增加修改头像接口
3.修改查询用户所有相关接口，需要加上用户头像的url返回，访问后可以直接获取到图片

说明：
注册的头像并不是和其他字段一起请求，而是需要1.先调用获取上传文件预签名接口，获取上传预签名地址，和文件唯一ID；2.通过预签名上传文件；3.请求注册接口，用户的头像字段使用第一步获取的文件ID。

任务3：
1.尝试实现角色权限管理，需要创建角色类，并且创建系统角色（1管理员；2普通用户）
2.所有用户都需要关联角色，并且需要创建权限信息，角色需要关联对应权限
3.增加全局拦截器，用户调用接口的时候需要验证是否拥有权限，如果不符合需要返回403权限不足错误

系统角色以及权限说明：
1.系统管理员：拥有所有权限
2.普通用户：只可以调用和自己相关的几个接口，比如修改自己的密码，头像这种和一些公共的查询接口
3.部分权限需要特殊处理，比如注册接口，不应该进行权限校验

+++ src/main/java/com/example/demo/dto/user/PageUserDTO.java
@@ -13,7 +13,6 @@ import java.util.Date;
 
 @Data
 public class PageUserDTO {
-
     @Schema(description = "用户名")
     private String userName;
 
@@ -29,18 +28,17 @@ public class PageUserDTO {
     @Schema(description = "性别枚举对象WOMAN女，MAN男")
     private SexEnum sexEnum;
 
-
     @Schema(description = "账号状态枚举")
     private StatusEnum statusEnum;
 
     @Schema(description = "页码", defaultValue = "1",required = true)
     @NotNull(message = "页码为空")
     @Min(value = 1, message = "页码必须大于0")
-    private Long page;
+    private int page;
 
     @Schema(description = "每页显示记录数", defaultValue = "10",required = true)
     @NotNull(message = "页大小为空")
     @Min(value = 1, message = "页大小不能小于1")
     @Max(value = 50, message = "页大小不能超过50")
-    private Long pageSize;
+    private int pageSize;
 }
